refactor: 불필요한 코드 삭제 및 미사용 파라미터 수정
