+++ src/main/java/nextstep/domain/QuerySetting.java
