+++ src/main/java/nextstep/domain/reservation/Reservation.java
@@ -2,7 +2,6 @@ package nextstep.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.Objects;
 
 public class Reservation {
     private Long id;
@@ -25,14 +24,6 @@ public class Reservation {
         this.time = time;
         this.name = name;
         this.theme = theme;
-    }
-
-    public boolean isSameDateAndTime(LocalDate date, LocalTime time) {
-        return this.date.equals(date) && this.time.equals(time);
-    }
-
-    public boolean isSameId(Long id) {
-        return Objects.equals(this.id, id);
     }
 
     public Long getId() {
